concesionario/requirements.txt
Django==1.8
Pillow==2.9.0
argparse==1.2.1
dj-database-url==0.4.0
dj-static==0.0.6
django-appconf==1.0.1
django-bootstrap3==6.2.2
django-cors-headers==1.1.0
django-easy-pdf==0.1.0
django-imagekit==3.2.7
django-toolbelt==0.0.1
djangorestframework==3.3.2
drf-nested-routers==0.11.1
gunicorn==19.4.5
pilkit==1.1.13
psycopg2==2.6.1
six==1.9.0
static3==0.6.1
wheel==0.24.0
wsgiref==0.1.2
